pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "numpy>=2.3.2",
    "onnxruntime>=1.22.1",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "uvicorn>=0.35.0",
]